rename distributed insertion file
